requirements.txt
pillow
psutil
requests
